feat: Introduce a dedicated experience page and refactor the resume root to display skills.
